add tests for methods in user service

+++ src/main/java/com/reservation_restaurants_service/entity/Restaurant.java
@@ -21,5 +21,5 @@ public class Restaurant {
     private String name;
     private String address;
     private long phoneNumber;
-    private double rating;
+    private long rating;
 }
